+++ R/zuxlsx-package.R
@@ -2,6 +2,8 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+## No .fixes: the entry points are registered as C_xlsx_sheets and
+## C_zuxlsx_native already, and .fixes = "C_" would prefix those again.
 #' @useDynLib zuxlsx, .registration = TRUE
 ## usethis namespace: end
 NULL